requirements.txt

# Core Framework and Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.10
alembic>=1.13.0

# Data Validation and Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6


# Machine Learning and Image Processing
torch>=2.2.0
torchvision>=0.17.0
opencv-python-headless>=4.8.1.78
numpy>=1.26.0
scikit-learn>=1.4.0
scipy>=1.11.0
pillow>=11.0.0
scikit-image>=0.22.0

# Development and Utilities
python-dotenv==1.0.0
httpx==0.25.2
psutil==5.9.5

# Optional Visualization
matplotlib>=3.10.0
